+++ src/utils/global-typescript.js
@@ -2,7 +2,7 @@
 
 module.exports = {
   target: 'es6',
-  module: 'es6',
+  module: 'commonjs',
   moduleResolution: 'node',
   declaration: true,
   experimentalDecorators: true