+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 - Improve ```mocha``` task
     - Improve parallel execution
+- Fix ```typescript``` transpile
+    - Support all import/export syntax of typescript e.g. export * from
 
 <a name="1.2.1"></a>
 ## [1.2.0](https://github.com/urbanjs/tools/compare/1.2.0...1.2.1) (2016-10-21)