+++ CHANGELOG.md
@@ -1,11 +1,12 @@
-## Unreleased
+<a name="1.2.3"></a>
+## [1.2.3](https://github.com/urbanjs/tools/compare/1.2.0...1.2.3) (2016-11-07)
 - Improve ```mocha``` task
     - Improve parallel execution
 - Fix ```typescript``` transpile
     - Support all import/export syntax of typescript e.g. export * from
 
 <a name="1.2.1"></a>
-## [1.2.0](https://github.com/urbanjs/tools/compare/1.2.0...1.2.1) (2016-10-21)
+## [1.2.1](https://github.com/urbanjs/tools/compare/1.2.0...1.2.1) (2016-10-21)
 - Fix ```mocha``` task
     - Fix parallel execution algorithm
     - Fix message intermixing